added initial Junit test class, changed History to have Double in get method instead of double so that null could be returned. Also added comments for jordan in the History tests

+++ src/exercise3/HistoryInterface.java
@@ -3,8 +3,8 @@ package exercise3;
 public interface HistoryInterface {
 	
 	public void clear();
-	public double get(int index);
-	public double add(double result);
+	public Double get(int index);
+	public int add(double result);
 	public double remove(int index);
 	public void printAll();
 }
